+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.1 - 2020-04-02
+
+### Fixed
+- `rgbdmux` - Fixed EOS forwarding by implementing custom `sink_event()` handling.
+- `rgbdmux` - No longer tries to renegotiate CAPS once EOS was received and the element is shutting down.
+
 ## 0.3.0 - 2020-03-10
 
 ### Added